% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_moon_image_latlong.R
\name{altaz_to_enu}
\alias{altaz_to_enu}
\title{Convert altitude/azimuth to ENU}
\usage{
altaz_to_enu(alt_rad, azN_rad)
}
\arguments{
\item{alt_rad}{Altitude in radians.}

\item{azN_rad}{Azimuth in radians measured from north.}
}
\description{
Map altitude/azimuth angles (north-referenced) to a unit
East-North-Up direction vector.
}
\keyword{internal}
