% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prague_rgb_correction.R
\name{apply_prague_rgb_gain}
\alias{apply_prague_rgb_gain}
\title{Apply Prague RGB correction gain}
\usage{
apply_prague_rgb_gain(x, gain)
}
\arguments{
\item{x}{RGB array or matrix.}

\item{gain}{RGB correction gain.}
}
\description{
Apply Prague RGB correction gain
}
\keyword{internal}
