% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{as_sky_image}
\alias{as_sky_image}
\title{Attach skymodelr RGB color metadata}
\usage{
as_sky_image(image)
}
\arguments{
\item{image}{Image array or rayimage image.}
}
\description{
Attach skymodelr RGB color metadata
}
\keyword{internal}
