% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars_radiometry.R
\name{bv_to_temperature_K}
\alias{bv_to_temperature_K}
\title{Convert B-V color index to blackbody temperature}
\usage{
bv_to_temperature_K(bv)
}
\arguments{
\item{bv}{Numeric vector of B-V values.}
}
\value{
Numeric vector of temperatures in Kelvin.
}
\description{
Convert B-V color index to blackbody temperature
}
\keyword{internal}
