% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiometry.R
\name{compute_spd_rgb_unit}
\alias{compute_spd_rgb_unit}
\title{Compute a unit RGB vector for a chosen SPD}
\usage{
compute_spd_rgb_unit(spd_type = c("D65", "BB5778"))
}
\arguments{
\item{spd_type}{SPD shape. Either \code{"D65"} or \code{"BB5778"}.}
}
\value{
Numeric vector length 3, normalized so sum equals 1.
}
\description{
Compute a unit RGB vector for a chosen SPD
}
