% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_moon_image_latlong.R
\name{cross_prod}
\alias{cross_prod}
\title{3D cross product helper}
\usage{
cross_prod(x, y)
}
\arguments{
\item{x}{Numeric vector of length 3.}

\item{y}{Numeric vector of length 3.}
}
\description{
Compute the 3D cross product of two numeric vectors.
}
\keyword{internal}
