% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{filter_supported_exr_metadata}
\alias{filter_supported_exr_metadata}
\title{Keep EXR metadata fields supported by libopenexr}
\usage{
filter_supported_exr_metadata(metadata)
}
\arguments{
\item{metadata}{EXR metadata list.}
}
\description{
Keep EXR metadata fields supported by libopenexr
}
\keyword{internal}
