% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{get_skymodelr_adopted_white}
\alias{get_skymodelr_adopted_white}
\title{Get skymodelr EXR adopted white}
\usage{
get_skymodelr_adopted_white(exr_adopted_white)
}
\arguments{
\item{exr_adopted_white}{Adopted white name or XYZ vector.}
}
\description{
Get skymodelr EXR adopted white
}
\keyword{internal}
