% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{jd_utc}
\alias{jd_utc}
\title{Convert POSIXct timestamps to Julian Date}
\usage{
jd_utc(time_utc)
}
\arguments{
\item{time_utc}{POSIXct vector in UTC.}
}
\description{
Transform a POSIXct time value into an astronomical Julian Date
(days since noon UTC 1 Jan 4713 BCE).
}
\keyword{internal}
