% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiometry.R
\name{lux_to_radiometric_irradiance}
\alias{lux_to_radiometric_irradiance}
\title{Convert photometric illuminance to radiometric irradiance}
\usage{
lux_to_radiometric_irradiance(E_v_lux, K_eff)
}
\arguments{
\item{E_v_lux}{Illuminance in lux.}

\item{K_eff}{Effective luminous efficacy in lm/W.}
}
\value{
Irradiance in W/m^2.
}
\description{
Convert photometric illuminance to radiometric irradiance
}
\keyword{internal}
