% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_starfield_rcpp}
\alias{make_starfield_rcpp}
\title{Render a star-field equirectangular EXR}
\usage{
make_starfield_rcpp(
  outfile,
  stars,
  resolution = 2048L,
  zero_point = 1,
  lon_deg = 0,
  lat_deg = 0,
  jd = 2451545,
  turbidity = 3,
  ozone_du = 300,
  altitude = 0,
  use_rgb = TRUE,
  numbercores = 1L
)
}
\arguments{
\item{outfile}{Output `.exr` path.}

\item{stars}{Data frame produced by `read_bsc5()` (must contain
`ra_rad`, `dec_rad`, `v_mag`).}

\item{resolution}{Horizontal *half-resolution*; final size is
2 × resolution × resolution.}

\item{zero_point}{Default `1`. Exposure scaling (see docs).}

\item{numbercores}{Number of CPU threads.}
}
\description{
Render a star-field equirectangular EXR
}
