% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makesky.R
\name{makesky}
\alias{makesky}
\title{Write a Hosek–Wilkie sky dome to EXR}
\usage{
makesky(
  outfile,
  albedo = 0.5,
  turbidity = 3,
  elevation = 10,
  resolution = 2048,
  numbercores = 1,
  square_projection = FALSE
)
}
\arguments{
\item{outfile}{Default `""`. Path to `.exr` file.}

\item{albedo}{Default `0.5`. 0–1 ground albedo.}

\item{turbidity}{Default `3`. 1.7–10 atmospheric turbidity.}

\item{elevation}{Default `10`. Solar elevation above the horizon (°).}

\item{resolution}{Default `2048`. Linear resolution of the square map (≥ 1).}

\item{numbercores}{Default `1`. Threads for \code{RcppThread}.}

\item{square_projection}{Default `FALSE`. If \code{TRUE} use equal‑area square mapping,
else latitude–longitude square mapping.}
}
\value{
Invisible `NULL`.  The EXR is written to `outfile`.
}
\description{
Write a Hosek–Wilkie sky dome to EXR
}
\examples{
makesky("sky.exr", turbidity = 2.4, elevation = 6, numbercores = 1)
}
