% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{makestars}
\alias{makestars}
\title{Write a star‑field EXR aligned with `makesky()`}
\usage{
makestars(
  outfile = "stars.exr",
  stars = NULL,
  resolution = 2048,
  zero_point = 1,
  lon_deg = 0,
  lat_deg = 0,
  time_utc = as.POSIXct("2000-01-01 00:00:00", tz = "UTC"),
  color = TRUE,
  ncores = parallel::detectCores()
)
}
\arguments{
\item{outfile}{`.exr` output path.}

\item{stars}{Tibble from `read_bsc5()`.  `NULL` ➜ packaged version.}

\item{resolution}{Map half‑width (image is 2×`resolution` × `resolution`).}

\item{zero_point}{Exposure scale (larger = brighter stars).}

\item{lon_deg}{Observer longitude (east +)}

\item{lat_deg}{Observer latitude (deg).}

\item{time_utc}{`POSIXct` UTC time stamp used to compute sidereal time.}

\item{ncores}{Threads for `RcppThread`.}
}
\description{
Write a star‑field EXR aligned with `makesky()`
}
