% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_moon_image_latlong.R
\name{moon_mag}
\alias{moon_mag}
\title{Moon V-band magnitude from phase angle}
\usage{
moon_mag(phase_deg, dist_km = 384400, mean_km = 384400)
}
\arguments{
\item{phase_deg}{Phase angle psi in degrees (0 = full Moon, 180 = new).}

\item{dist_km}{Default \code{384400}. Current geocentric distance (km).}

\item{mean_km}{Default \code{384400}. Mean distance (km).}
}
\value{
Apparent V magnitude \code{m}.
}
\description{
Moon V-band magnitude from phase angle
}
\keyword{internal}
