% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiometry.R
\name{moon_solid_angle_sr}
\alias{moon_solid_angle_sr}
\title{Moon disk solid angle in steradians}
\usage{
moon_solid_angle_sr(diameter_deg = 0.533)
}
\arguments{
\item{diameter_deg}{Angular diameter in degrees.}
}
\value{
Solid angle in steradians.
}
\description{
Moon disk solid angle in steradians
}
\keyword{internal}
