% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prague_rgb_correction.R
\name{normalize_prague_rgb_correction}
\alias{normalize_prague_rgb_correction}
\title{Normalize Prague RGB correction option}
\usage{
normalize_prague_rgb_correction(prague_rgb_correction)
}
\arguments{
\item{prague_rgb_correction}{Prague RGB correction option.}
}
\description{
Normalize Prague RGB correction option
}
\keyword{internal}
