% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makesky.R
\name{raw_prague}
\alias{raw_prague}
\title{Write a Hosek–Wilkie sky dome to EXR}
\usage{
raw_prague(
  phi,
  theta,
  altitude = 0,
  elevation = 10,
  visibility = 50,
  albedo = 0.5,
  azimuth = 90,
  resolution = 2048,
  numbercores = 1,
  wide_spectrum = FALSE
)
}
\arguments{
\item{altitude}{Default `0`. Altitude of the viewer in meters. Valid [0,15000]. Only valid for the
Prague model.}

\item{elevation}{Default `10`. Solar elevation above the horizon (°).}

\item{visibility}{Default `50`. Meteorological range in kilometres for Prague model.}

\item{albedo}{Default `0.5`. 0–1 ground albedo.}

\item{azimuth}{Default `90`. Solar azimuth (°). Defaults South.}

\item{resolution}{Default `2048`. Height of the image. Width is 2x this number.}

\item{numbercores}{Default `1`. Number of threads to use in computation.}

\item{outfile}{Default `""`. Path to `.exr` file.}

\item{turbidity}{Default `3`. 1.7–10 atmospheric turbidity. Only valid for Hosek model.}

\item{square_projection}{Default `FALSE`. If \code{TRUE} use equal‑area square mapping,
else latitude–longitude square mapping.}

\item{hosek}{Default `TRUE`. Use `"prague"` to enable the Prague 2021-22 spectral sky model.}

\item{prg_dataset}{Default `""`. Full path to the Prague binary dataset (`*.dat`) when `model = "prague"`.}
}
\value{
Invisible `NULL`.  The EXR is written to `outfile`.
}
\description{
Write a Hosek–Wilkie sky dome to EXR
}
\examples{
#Generate a basic atmosphere with the Hosek model
temp_exr = tempfile(fileext=".exr")
generate_sky(temp_exr, turbidity = 3, elevation = 2, numbercores = 1)
}
