% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{sky_exr_metadata}
\alias{sky_exr_metadata}
\title{Build EXR metadata for skymodelr RGB output}
\usage{
sky_exr_metadata(image = NULL)
}
\arguments{
\item{image}{Default \code{NULL}. Image to read colorspace and white metadata from.}
}
\description{
Build EXR metadata for skymodelr RGB output
}
\keyword{internal}
