% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_moon_image_latlong.R
\name{swe_dirs_topo_planets_df}
\alias{swe_dirs_topo_planets_df}
\title{Compute bright-planet positions via Swiss Ephemeris}
\usage{
swe_dirs_topo_planets_df(datetime, lat, lon, elev_m = 0)
}
\arguments{
\item{datetime}{POSIXct observation time (UTC).}

\item{lat}{Latitude in degrees.}

\item{lon}{Longitude in degrees.}

\item{elev_m}{Observer elevation above sea level in metres.}
}
\description{
Produce a data frame of planetary apparent positions and
magnitudes for the observer.
}
\keyword{internal}
