% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_moon_image_latlong.R
\name{sweph_info_to_stars_df}
\alias{sweph_info_to_stars_df}
\title{Convert Swiss ephemeris data to a starfield frame}
\usage{
sweph_info_to_stars_df(sweph_info)
}
\arguments{
\item{sweph_info}{List of planetary info objects.}
}
\description{
Build a data frame compatible with `make_starfield_rcpp()` from
Swiss Ephemeris results.
}
\keyword{internal}
