% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{tag_generated_sky_exr_metadata}
\alias{tag_generated_sky_exr_metadata}
\title{Tag generated sky EXR metadata}
\usage{
tag_generated_sky_exr_metadata(sky, hosek, exr_metadata, exr_adopted_white)
}
\arguments{
\item{sky}{Sky image array.}

\item{hosek}{Whether the Hosek-Wilkie model was used.}

\item{exr_metadata}{Whether to attach skymodelr EXR metadata.}

\item{exr_adopted_white}{Adopted white name or XYZ vector.}
}
\description{
Tag generated sky EXR metadata
}
\keyword{internal}
