% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{tag_skymodelr_exr_metadata}
\alias{tag_skymodelr_exr_metadata}
\title{Tag skymodelr EXR metadata}
\usage{
tag_skymodelr_exr_metadata(
  sky,
  adopted_white_xy = .skymodelr_d60_xy,
  adopted_white_name = .skymodelr_exr_adopted_neutral_label,
  rgb_colorspace = rayimage::CS_SRGB,
  model_name = NULL,
  prague_rgb_correction = NULL
)
}
\arguments{
\item{sky}{Sky image array.}

\item{adopted_white_xy}{Default \code{.skymodelr_d60_xy}. Adopted white xy.}

\item{adopted_white_name}{Default \code{.skymodelr_exr_adopted_neutral_label}.
Adopted white name.}

\item{rgb_colorspace}{Default \code{rayimage::CS_SRGB}. RGB colorspace.}

\item{model_name}{Default \code{NULL}. Sky model name.}

\item{prague_rgb_correction}{Default \code{NULL}. Prague RGB correction label.}
}
\description{
Tag skymodelr EXR metadata
}
\keyword{internal}
