% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prague_rgb_correction.R
\name{validate_prague_rgb_gain}
\alias{validate_prague_rgb_gain}
\title{Validate Prague RGB correction gain}
\usage{
validate_prague_rgb_gain(gain)
}
\arguments{
\item{gain}{RGB correction gain.}
}
\description{
Validate Prague RGB correction gain
}
\keyword{internal}
