% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars_radiometry.R
\name{vmag_to_ev_lux}
\alias{vmag_to_ev_lux}
\title{Convert V magnitude to illuminance in lux}
\usage{
vmag_to_ev_lux(mV, zero_point = 1)
}
\arguments{
\item{mV}{Numeric vector of V magnitudes.}

\item{zero_point}{Optional scalar multiplier for catalog scaling.}
}
\value{
Numeric vector of photometric illuminance in lux.
}
\description{
Convert V magnitude to illuminance in lux
}
\keyword{internal}
