% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{write_sky_image}
\alias{write_sky_image}
\title{Write a skymodelr image with EXR color metadata}
\usage{
write_sky_image(image, filename, ...)
}
\arguments{
\item{image}{Image array or rayimage image.}

\item{filename}{Destination image path.}

\item{...}{Additional arguments passed to \code{\link[rayimage:ray_write_image]{rayimage::ray_write_image()}}.}
}
\description{
Write a skymodelr image with EXR color metadata
}
\keyword{internal}
