% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{xy_to_xyz_y1}
\alias{xy_to_xyz_y1}
\title{Convert xy chromaticity to XYZ with Y equal to 1}
\usage{
xy_to_xyz_y1(xy)
}
\arguments{
\item{xy}{Finite xy chromaticity vector.}
}
\description{
Convert xy chromaticity to XYZ with Y equal to 1
}
\keyword{internal}
