% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sky.R
\name{calculate_sky_values}
\alias{calculate_sky_values}
\title{Sample a direction from the Prague model.}
\usage{
calculate_sky_values(
  phi,
  theta,
  altitude = 0,
  elevation = 10,
  visibility = 50,
  albedo = 0.5,
  azimuth = 90,
  numbercores = 1,
  wide_spectrum = FALSE
)
}
\arguments{
\item{phi}{Horizontal angle of the sample, degrees. Vectorized. Range [0, 360].}

\item{theta}{Vertical angle of the sample, degrees. Vectorized. Range [-90, 90].}

\item{altitude}{Default `0`, vectorized. Altitude of the viewer in meters. Range [0, 15000].}

\item{elevation}{Default `10`, vectorized. Solar elevation angle above/below the horizon (degrees). Range [-4.2, 90].}

\item{visibility}{Default `50`, vectorized. Range [20, 131.8]. Meteorological range in kilometers for Prague model.}

\item{albedo}{Default `0.5`, vectorized. Range [0, 1]. Ground albedo.}

\item{azimuth}{Default `90`, single value. Solar azimuth (degrees). Defaults South.}

\item{numbercores}{Default `1`. Number of threads to use in computation.}

\item{wide_spectrum}{Default `FALSE`. Whether to use the wide‑spectrum (55‑channel, polarised) coefficients.}
}
\value{
3 column RGB matrix.
}
\description{
Sample a direction from the Prague model.
}
\examples{
#Generate a basic atmosphere with the Hosek model
if(run_documentation()) {
value_grid = expand.grid(
  phi = seq(0,360,by=30),
  theta = seq(0,90,by=10),
  altitude = c(0,10000)
 )
vals = calculate_sky_values(
  phi = value_grid$phi,
  theta = value_grid$theta,
  altitude = value_grid$altitude,
  elevation = 45,
  visibility = 120,
  albedo = 0
 )
 cbind(value_grid, vals)
}
}
