% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_moon_image_latlong.R
\name{generate_moon_image_latlong}
\alias{generate_moon_image_latlong}
\title{Render the moon into a lat-long EXR patch}
\usage{
generate_moon_image_latlong(
  datetime,
  lat,
  lon,
  elev_m = 0,
  width = 400,
  height = 400
)
}
\arguments{
\item{datetime}{POSIXct observation time (UTC).}

\item{lat}{Latitude in degrees.}

\item{lon}{Longitude in degrees.}

\item{elev_m}{Observer elevation above sea level in metres.}

\item{width}{Output width in pixels.}

\item{height}{Output height in pixels.}
}
\description{
Produce a rasterised moon texture aligned for composition into
the sky dome.
}
\keyword{internal}
