% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{generate_moon_latlong}
\alias{generate_moon_latlong}
\title{Generate the atmosphere with the moon}
\usage{
generate_moon_latlong(
  outfile = NA,
  datetime = "2025-07-29 18:00:00",
  lat = 38.9072,
  lon = -77.0369,
  albedo = 0.5,
  turbidity = 3,
  altitude = 0,
  resolution = 2048,
  numbercores = 1,
  hosek = TRUE,
  wide_spectrum = FALSE,
  visibility = 50,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{outfile}{Default `NA`. Path to the `.exr` file to write.}

\item{datetime}{Default `"2025-07-29 18:00:00"`. POSIX-compatible
date-time; if missing a time-zone it is assumed to be local.}

\item{lat}{Default `38.9072`. Observer latitude (degrees N).}

\item{lon}{Default `-77.0369`. Observer longitude (degrees E; west < 0).}

\item{albedo}{Default `0.5`. Ground albedo, range [0, 1].}

\item{turbidity}{Default `3`. Atmospheric turbidity, range
[1.7, 10] (*Hosek only*).}

\item{altitude}{Default `0`. Observer altitude (m), range
[0, 15000] (*Prague only*).}

\item{resolution}{Default `2048`. Image height in pixels (width = 2 × height).}

\item{numbercores}{Default `1`. CPU threads to use.}

\item{hosek}{Default `TRUE`. `FALSE` selects the Prague model.}

\item{wide_spectrum}{Default `FALSE`. 55-channel Prague coefficients (altitude = 0 m only).}

\item{visibility}{Default `50`. Meteorological range (km); *Prague only*.}

\item{verbose}{Default `FALSE`. Whether to print progress bars/diagnostic info.}

\item{...}{Additional arguments passed to [generate_stars()]}
}
\value{
Either the raw data, or the data is invisibly returned if outfile is given. The EXR is written to `outfile`.
}
\description{
Note that this is just a scaled version of `generate_sky()`, scaled down by the luminance
of the moon as compared to the sun. This function takes the phase of the moon into account,
along with the increase in luminosity around a full moon (known as opposition surge).
}
\examples{
# Moonlit sky (Hosek), mid-evening in DC
if(run_documentation()) {
generate_moon_latlong(
  datetime   = as.POSIXct("2025-07-30 20:20:00",tz="EST"),
  lat        = 38.9072,
  lon        = -77.0369,
  resolution = 400,
  turbidity  = 3,
  verbose    = TRUE
) |>
  rayimage::render_exposure(15) |>
  rayimage::plot_image()
}
}
