% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sky.R
\name{generate_sky}
\alias{generate_sky}
\title{Write a Hosek–Wilkie sky dome to EXR}
\usage{
generate_sky(
  outfile = NA,
  albedo = 0.5,
  turbidity = 3,
  elevation = 10,
  azimuth = 90,
  altitude = 0,
  resolution = 2048,
  numbercores = 1,
  hosek = TRUE,
  wide_spectrum = FALSE,
  visibility = 50,
  square_projection = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{outfile}{Default `NA`. Path to `.exr` file. If not given, the data will be returned instead.}

\item{albedo}{Default `0.5`. 0–1 ground albedo.}

\item{turbidity}{Default `3`. 1.7–10 atmospheric turbidity. Only valid for Hosek model.}

\item{elevation}{Default `10`. Solar elevation above the horizon (degrees).}

\item{azimuth}{Default `90`. Solar azimuth (degrees). Defaults South.}

\item{altitude}{Default `0`. Altitude of the viewer in meters. Valid [0,15000]. Only valid for the
Prague model.}

\item{resolution}{Default `2048`. Height of the image. Width is 2x this number.}

\item{numbercores}{Default `1`. Number of threads to use in computation.}

\item{hosek}{Default `TRUE`. Use `"prague"` to enable the Prague 2021-22 spectral sky model.}

\item{visibility}{Default `50`. Meteorological range in kilometres for Prague model.}

\item{square_projection}{Default `FALSE`. If \code{TRUE} use equal‑area square mapping,
else latitude–longitude square mapping.}

\item{verbose}{Default `FALSE`. Whether to print progress bars/diagnostic info.}
}
\value{
Either the raw data, or the data is invisibly returned if outfile is given. The EXR is written to `outfile`.
}
\description{
Write a Hosek–Wilkie sky dome to EXR
}
\examples{
#Generate a basic atmosphere with the Hosek model
temp_exr = tempfile(fileext=".exr")
generate_sky(temp_exr, turbidity = 3, elevation = 2, numbercores = 1)
}
