% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{generate_stars}
\alias{generate_stars}
\title{Write a star‑field EXR aligned with `generate_sky()`}
\usage{
generate_stars(
  outfile = "stars.exr",
  resolution = 2048,
  zero_point = 1,
  lon_deg = 0,
  lat_deg = 0,
  time_utc = as.POSIXct("2000-01-01 00:00:00", tz = "UTC"),
  turbidity = 3,
  ozone_du = 300,
  altitude = 0,
  color = TRUE,
  ncores = parallel::detectCores()
)
}
\arguments{
\item{outfile}{`.exr` output path.}

\item{resolution}{Map half‑width (image is 2×`resolution` × `resolution`).}

\item{zero_point}{Exposure scale (larger = brighter stars).}

\item{lon_deg}{Observer longitude (east +)}

\item{lat_deg}{Observer latitude (deg).}

\item{time_utc}{`POSIXct` UTC time stamp used to compute sidereal time.}

\item{ncores}{Number of threads to use in computation.}
}
\description{
Write a star‑field EXR aligned with `generate_sky()`
}
