% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makesky.R
\name{makesky}
\alias{makesky}
\title{Write a Hosek–Wilkie sky dome to EXR}
\usage{
makesky(
  outfile,
  albedo = 0.5,
  turbidity = 3,
  elevation = 10,
  azimuth = 90,
  altitude = 0,
  resolution = 2048,
  numbercores = 1,
  model = "hosek",
  prg_dataset = "",
  visibility = 50,
  square_projection = FALSE
)
}
\arguments{
\item{outfile}{Default `""`. Path to `.exr` file.}

\item{albedo}{Default `0.5`. 0–1 ground albedo.}

\item{turbidity}{Default `3`. 1.7–10 atmospheric turbidity.}

\item{elevation}{Default `10`. Solar elevation above the horizon (°).}

\item{azimuth}{Default `90`. Solar azimuth (°). Defaults South.}

\item{altitude}{Default `0`. Altitude of the viewer in meters. Valid [0,15000]. Only valid for the
Prague model.}

\item{resolution}{Default `2048`. Linear resolution of the square map (≥ 1).}

\item{numbercores}{Default `1`. Threads for \code{RcppThread}.}

\item{model}{Default `"hosek"`. Use `"prague"` to enable the Prague 2021-22 spectral sky model.}

\item{prg_dataset}{Default `""`. Full path to the Prague binary dataset (`*.dat`) when `model = "prague"`.}

\item{visibility}{Default `50`. Meteorological range in kilometres for Prague model.}

\item{square_projection}{Default `FALSE`. If \code{TRUE} use equal‑area square mapping,
else latitude–longitude square mapping.}
}
\value{
Invisible `NULL`.  The EXR is written to `outfile`.
}
\description{
Write a Hosek–Wilkie sky dome to EXR
}
\examples{
makesky("sky.exr", turbidity = 2.4, elevation = 6, numbercores = 1)
}
