% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prague_rgb_correction.R
\name{prepare_prague_rgb_gain}
\alias{prepare_prague_rgb_gain}
\title{Prepare Prague RGB correction gain}
\usage{
prepare_prague_rgb_gain(gain = .prague_rgb_correction_gain, strength = 1)
}
\arguments{
\item{gain}{Default \code{.prague_rgb_correction_gain}. RGB correction gain.}

\item{strength}{Default \code{1}. Correction strength.}
}
\description{
Prepare Prague RGB correction gain
}
\keyword{internal}
