% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_moon_image_latlong.R
\name{swe_dirs_topo_moon_sun}
\alias{swe_dirs_topo_moon_sun}
\title{Compute topocentric moon/sun directions}
\usage{
swe_dirs_topo_moon_sun(datetime, lat, lon, elev_m = 0)
}
\arguments{
\item{datetime}{POSIXct observation time (UTC).}

\item{lat}{Latitude in degrees.}

\item{lon}{Longitude in degrees.}

\item{elev_m}{Observer elevation above sea level in metres.}
}
\description{
Query Swiss Ephemeris for topocentric sun and moon direction
vectors and magnitudes for a given observer location.
}
\keyword{internal}
